some global variables moved from MovieCenter.py to CommonSupport.py to avoid cross import references
