+++ src/MovieSelection.py
@@ -73,10 +73,9 @@ from .MovieRetitle import MovieRetitle
 from Components.Sources.EMCServiceEvent import EMCServiceEvent
 from .MovieInfo import MovieInfoTMDb
 from .EMCPlayList import emcplaylist, EMCPlaylistScreen, EMCPlaylistSetup
-from .CommonSupport import getInfoFile
+from .CommonSupport import getInfoFile, extList, extVideo, extMedia, extDir, extPlaylist, extAudio, plyAll, plyDVD, plyVideo, plyDVB
 
-from .MovieCenter import extList, extVideo, extMedia, extDir, plyAll, plyDVD, plyVideo, cmtBME2, cmtBMEMC, cmtDir, plyDVB, extPlaylist, extAudio
-from .MovieCenter import getMovieNameWithoutExt, getMovieNameWithoutPhrases, getNoPosterPath, getPosterPath
+from .MovieCenter import getMovieNameWithoutExt, getMovieNameWithoutPhrases, getNoPosterPath, getPosterPath, cmtBME2, cmtBMEMC, cmtDir
 
 from six.moves import range
