+++ src/MovieSelectionMenu.py
@@ -45,7 +45,7 @@ from .PermanentSort import PermanentSort
 from .E2Bookmarks import E2Bookmarks
 from .EMCBookmarks import EMCBookmarks
 from .RogueFileCheck import RogueFileCheck
-from .MovieCenter import extTS, extMedia
+from .CommonSupport import extTS, extMedia
 global extTS
 
 cutsParser = struct.Struct('>QI') # big-endian, 64-bit PTS and 32-bit type
